Test implementation of pulling the Reader out of the ZipArchive structure.
It appears to work!  But is not terribly convenient.  In particular, it's
entirely possible to give a ZipFile the wrong reader, and the ZipFile only
borrows the reader which is not always what you want.  (Namely, not really
what I want.)

Going to experiment with the ownership semantics of this more; it might
be easiest/best(??) to just slap the Reader into an Arc<Refcell<_>> or
such and call it a day.