Made ZipFile own its Reader.
Ugh, it's pretty nasty.  I think the RIGHT way might be to separate
out owned-readers and borrowed-readers so you can have an API that
just borrows stuff when necessary.  Either that or deal with locking
the file and seeking back and forth in it when one of the file readers
is reading from it.

An into_inner() method to recover the Reader from a ZipFile might also
make it nicer when you want to read multiple things from the same Read.

What is now the ZipArchive could become a ZipHeader and be shared among
multiple things.  For now we kinda clone the relevant parts.

Maybe we have the ZipArchive which contains metadata for the file,
which is only scanned once, then a ZipReader that consumes a particular
Read but might let you read multiple ZipFiles from it?  That might
work.